refactor: move SandRenderer to Playground folder
